add dry run arg

+++ engagement_database/delete_datasets.py
@@ -24,6 +24,8 @@ class keyvalue(argparse.Action):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Deletes messages for given datasets in engagement database")
 
+    parser.add_argument("--dry-run", action="store_true",
+                        help="Logs the updates that would be made without updating anything.")
     parser.add_argument("google_cloud_credentials_file_path", metavar="google-cloud-credentials-file-path",
                         help="Path to a Google Cloud service account credentials file to use to access the "
                              "credentials bucket")
@@ -39,11 +41,15 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
+    dry_run = args.dry_run
     google_cloud_credentials_file_path = args.google_cloud_credentials_file_path
     engagement_database_credentials_file_url = args.engagement_database_credentials_file_url
     database_path = args.database_path
     coda_credentials_file_url = args.coda_credentials_file_url
     engagement_db_to_coda_dataset = args.engagement_db_to_coda_dataset
+
+    dry_run_text = "(dry run)" if dry_run else ""
+    log.info(f"Deletes messages in given engagement database {dry_run_text}")
 
     log.info("Downloading Firestore engagement database credentials...")
     engagement_database_credentials = json.loads(google_cloud_utils.download_blob_to_string(
@@ -68,13 +74,16 @@ if __name__ == "__main__":
             if msg.coda_id is not None:
                 coda_message = coda.get_dataset_message(coda_dataset, msg.coda_id)
                 assert coda_message is not None
-                coda.delete_dataset_message(coda_dataset, msg.coda_id)
+                if not dry_run:
+                    coda.delete_dataset_message(coda_dataset, msg.coda_id)
 
             log.warning(f"Deleting engagement db message with message id {msg.message_id}")
-            engagement_db.delete_doc(f"messages/{msg.message_id}")
+            if not dry_run:
+                engagement_db.delete_doc(f"messages/{msg.message_id}")
 
             history_entries = engagement_db.get_history_for_message(msg.message_id)
             log.warning(f"Found {len(history_entries)} engagement db message's history entries")
             for history_entry in history_entries:
                 log.warning(f"Deleting history entry with id {history_entry.history_entry_id}")
-                engagement_db.delete_doc(f"history/{history_entry.history_entry_id}")
+                if not dry_run:
+                    engagement_db.delete_doc(f"history/{history_entry.history_entry_id}")
